last update

+++ lib/pages/globals.dart
@@ -2,7 +2,7 @@ library my_prj.globals;
 
 import 'package:ir_simulation/models/simulation_ir.dart';
 
-int nbrKids = SimulationIr.nbrKids;
+double nbrKids = SimulationIr.nbrKids;
 double baseSalary = SimulationIr.baseSalary;
 var detentionLit = SimulationIr.detentionList;
 var indemniteLit = SimulationIr.indemniteLit;
